refactor(homepage): done
