+++ src/components/styles/contact.module.css
@@ -1,30 +1,31 @@
 .contact {
-  background-color: var(--color-purple-faded);
+  background-color: var(--color-purple-light);
   border-radius: 37.3588px;
   display: grid;
   grid: 1fr 1fr / 1fr 1fr;
   grid-row-gap: 1rem;
-  grid-column-gap: 2rem;
+  grid-column-gap: 1rem;
 
   place-items: center;
-  padding: 1rem 1.5rem;
+  padding: 14px 32px;
 }
 
 .contact .icons a > svg {
-  width: 6rem;
-  height: 6rem;
-  padding: 0.5rem 1rem;
+  width: auto;
+  height: 2.25rem;
   box-sizing: border-box;
   display: inline;
 }
 
 .contact .brand {
   display: none;
+  height: 2.25rem;
+  width: auto;
 }
 
 .contact span {
   color: white;
-  font-size: var(--font-size-bigger);
+  font-size: var(--font-size-big);
   grid-row-start: 1;
   grid-row-end: 1;
   grid-column: 1 / span 2;
@@ -55,7 +56,6 @@
     grid-column: 1 / -1;
     justify-self: start;
     grid-column: 1;
-    height: 75%;
     grid-column: 1 / span 1;
     grid-row: 1 / span 3;
   }
@@ -75,7 +75,7 @@
 
 @media screen and (min-width: 850px) {
   .contact {
-    grid: 1fr / [logo] 4fr [text] 16fr [icons] 6fr;
+    grid: 1fr / [logo] auto [text] 16fr [icons] 6fr;
   }
 
   .contact .brand {
@@ -83,7 +83,6 @@
 
     display: block;
     grid-column: logo;
-    height: 100px;
   }
 
   .contact .rest {
