+++ src/pages/styles/index.module.css
@@ -199,12 +199,11 @@
 .index .partners svg {
   margin: 2rem;
   margin-top: 1rem;
-  max-width: 250px;
-  max-height: 60px;
+  max-width: 200px;
+  max-height: 80px;
 }
 
 .index .partners > h2 {
-  font-weight: var(--font-weight-bold);
 }
 
 .index :global(.slick-list) {
