+++ src/styles/_layout.scss
@@ -44,7 +44,6 @@ body {
 main {
   background: linear-gradient(316.86deg, #1e075f 56.13%, #4d00b4 97.68%);
   word-break: break-word;
-  padding-bottom: 4rem;
 }
 
 main > div {
